KClient Completed

+++ KClient.h
@@ -46,8 +46,8 @@ private:
     ifstream encryptedDataToStream(const Ciphertext &);
     vector<ZZ_pX> loadeddata;
     vector<ZZ_p> labels;
-    map<size_t,ZZ_pX > encrypted_data_hash_table;
-    map<size_t,unsigned> results;
+    map<uint32_t ,ZZ_pX > encrypted_data_hash_table;
+    map<uint32_t ,unsigned> results;
     void connectToUServer();
     void connectToTServer();
